3.architecture_pour_la_prediction/3.3.prediction_cloud_dataflow_prophet/requirements_before_template.txt
#google-apitools==0.5.31
#google-api-core==1.22.0
#google-apitools==0.5.31
#google-auth==1.31.0
#google-auth-oauthlib==0.4.4
#google-cloud-bigquery==1.26.1
#google-cloud-bigtable==1.0.0
#google-cloud-build==2.0.0
#google-cloud-core==1.4.1
#google-cloud-datastore==1.7.4
#google-cloud-dlp==1.0.0
#google-cloud-language==1.3.0
#google-cloud-pubsub==1.0.2
#google-cloud-spanner==1.19.1
#google-cloud-videointelligence==1.16.1
#google-cloud-vision==1.0.0
#google-crc32c==1.1.2
#google-pasta==0.2.0
#google-resumable-media==1.3.0
#googleapis-common-protos==1.53.0
#grpc-google-iam-v1==0.12.3
#grpcio==1.38.0
#grpcio-gcp==0.2.2
#pandas==1.2.4
#Cython==0.29.13
#setuptools-git==1.2
#apache-beam==2.25.0

apache-beam==2.25.0
absl-py==0.13.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
avro-python3==1.8.2
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cmdstanpy==0.9.68
convertdate==2.3.2
crcmod==1.7
cycler==0.10.0
Cython==0.29.13
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.1.1
docopt==0.6.2
entrypoints==0.3
ephem==4.0.0.2
facets-overview==1.0.0
fastavro==1.0.0.post1
fasteners==0.16.3
future==0.18.2
gast==0.3.3
google-api-core==1.22.0
google-apitools==0.5.31
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==1.26.1
google-cloud-bigtable==1.0.0
google-cloud-build==2.0.0
google-cloud-core==1.4.1
google-cloud-datastore==1.7.4
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
grpcio-gcp==0.2.2
guppy3==3.0.9
h5py==2.10.0
hdfs==2.5.8
hijri-converter==2.1.3
holidays==0.11.1
httplib2==0.12.0
idna==2.10
importlib-metadata==4.5.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
libcst==0.3.19
LunarCalendar==0.0.9
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
mmh3==2.5.1
mock==2.0.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
numpy==1.17.3
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.12.2
protorpc==0.12.0
ptyprocess==0.7.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.9.0
PyHamcrest==1.10.1
PyMeeus==0.5.11
pymongo==3.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
python-gflags==3.0.6
pytz==2019.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
Send2Trash==1.5.0
setuptools-git==1.2
six==1.16.0
sql==0.4.0
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
timeloop==1.0.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.7.1
ujson==4.0.2
urllib3==1.26.5
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1